asr-11-experiment/productos/requirements.txt
aniso8601==10.0.0
APScheduler==3.11.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
Flask-RESTful==0.3.10
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
psutil==7.0.0
py-healthcheck==1.10.1
pytz==2025.1
requests==2.32.3
six==1.17.0
tzlocal==5.3
urllib3==2.3.0
Werkzeug==3.1.3
gunicorn==20.1.0
flask_monitoringdashboard==3.3.2